requirements.txt
line-bot-sdk
flask
configparser
openai
azure-ai-translation-text